Springer Free Ebooks Downloader
This Shell's goal is to download all the Free Ebooks that the editor Springer made available during 2020's Covid-19 epidemic.
Source: https://group.springernature.com/gp/group/media/press-releases/freely-accessible-textbook-initiative-for-educators-and-students/17858180?utm_medium=social&utm_content=organic&utm_source=facebook&utm_campaign=SpringerNature_&sf232256230=1

Obs: It's the second version of it, that, instead of obtaining all the books via a news website, this script will act and download everything solely on the editor's website.
Also, it is meant to download all the books avaible.

It's also worth noting that, once this giveaway is over or if Springer make any changes on it's website structure, this script won't work anymore.
Also, this is meant merely by testing purposes and only for the free ebooks. All books' copywrite belongs to Springer and all the book authors.